% Generated by roxygen2 (4.0.1): do not edit by hand
\name{basic.stats}
\alias{basic.stats}
\title{Basic statistical estimators}
\usage{
basic.stats(data)
}
\arguments{
\item{data}{vector}
}
\value{
data.frame
}
\description{
\code{basic.stats} is a simple basic statistical estimators calculation
}
\details{

}
\author{
I. Lopez-Coto, 2012 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

