% Generated by roxygen2 (4.0.1): do not edit by hand
\name{color.themes}
\alias{color.themes}
\title{Color Themes}
\usage{
color.themes(theme, n, rev)
}
\arguments{
\item{theme}{\code{string} or \code{integer} indicating the desired color scheme c("radioactive","galaxyblue","bw",'pinkpanther',"iceblue",'apocalypse_now','lime','luna','bossanova','trideca')}

\item{n}{\code{integer} number of colors}

\item{rev}{\code{logical} reverse order}
}
\value{
a list with a vector of colors (colsch), and 3 independent colors (col1, col2, st.marker.col). In addition, load the library \link{RColorBrewer}
}
\description{
\code{color.themes} provides different color schemes
}
\details{

}
\author{
I. Lopez-Coto, 2014 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

