% Generated by roxygen2 (4.0.1): do not edit by hand
\name{filled.contour2}
\alias{filled.contour2}
\title{Plot color filled contours}
\usage{
filled.contour2(x = seq(0, 1, length.out = nrow(z)), y = seq(0, 1,
  length.out = ncol(z)), z, xlim = range(x, finite = TRUE), ylim = range(y,
  finite = TRUE), zlim = range(z, finite = TRUE), levels = pretty(zlim,
  nlevels), nlevels = 20, color.palette = cm.colors,
  col = color.palette(length(levels) - 1), plot.title, plot.axes, key.title,
  key.axes, asp = NA, xaxs = "i", yaxs = "i", las = 1, axes = TRUE,
  frame.plot = axes, mar, ...)
}
\arguments{
\item{see}{\link{filled.contour}}
}
\description{
\code{filled.contour2} is modification by Ian Taylor of the filled.contour function
 to remove the key and facilitate overplotting with contour(). Bug fixed by I. Lopez-Coto
}
\details{

}
\author{
Ian Taylor and I. Lopez-Coto, 2013 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

