% Generated by roxygen2 (4.0.1): do not edit by hand
\name{filledContour2}
\alias{filledContour2}
\title{Plot color filled contours}
\usage{
filledContour2(x, y = 1, maxpixels = 1e+05, ...)
}
\arguments{
\item{see}{\link{filled.contour}}
}
\description{
\code{filledContour2} is a wrapper to filled.contour2 with raster objects
}
\details{

}
\author{
I. Lopez-Coto, 2013 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

