% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getCONC4st}
\alias{getCONC4st}
\title{Extract WRF data at one station location}
\usage{
getCONC4st(filename, lon_WRF, lat_WRF, lon, lat, agl, var = c("CO2_BIO",
  "CO2_ANT"), vlev = 10)
}
\arguments{
\item{filename}{a string with the WRF filename}

\item{lon_WRF}{2D WRF longitudes matrix as produced by \link{getGRID}}

\item{lat_WRF}{2D WRF latitudes matrix as produced by \link{getGRID}}

\item{lon}{station longitude}

\item{lat}{station latitude}

\item{agl}{station height above ground level}

\item{var}{vector of strings containing the name(s) of 4D WRF variable(s) to extract}

\item{vlev}{number of vertical levels (default = 10)}
}
\value{
list containing objects, mostly rasters, for the variables \cr
 c('U3D','V3D','W3D','WS3D','WD3D','P3D','T3D','TH3D','TD3D',var4d,var3d)
}
\description{
\code{getCONC4st} extracts 4D WRF variables at ONE station location from a WRF file. The purpose of this function is to extract a time serie as fast as possible
}
\details{

}
\author{
I. Lopez-Coto, 2014 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

