% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getGRID}
\alias{getGRID}
\title{Get the WRF grid}
\usage{
getGRID(filename)
}
\arguments{
\item{filename}{input filename}
}
\value{
a list containing the lat and lon arrays and projection parameters
}
\description{
\code{getGRID} is a function to get the WRF grid and related projection parameters
}
\details{

}
\examples{
\dontrun{
 WRFgrid <- getGRID(paste(pathWRF,"wrfinput_",domain[d],sep=''))
 for (i in length(names(WRFgrid))){
 assign(names(WRFgrid)[[i]],WRFgrid[[i]])
 }
 }
}
\author{
I. Lopez-Coto, 2013 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

