% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getISD}
\alias{getISD}
\title{Retrieve the ISD data file}
\usage{
getISD(year = 2014, USAF = 84490, WBAN = 13025, stations = NULL)
}
\arguments{
\item{pathout}{}

\item{year}{4 digits year}

\item{USAF}{usaf station code (6 digits but it accepts less)}

\item{WBAN}{wban station code (5 digits but it accepts less)}
}
\value{
character vector with one element per raw ISD report
}
\description{
\code{getISD} download the station data from the ftp://ftp.ncdc.noaa.gov/pub/data/noaa/...
}
\details{

}
\author{
I. Lopez-Coto, 2013 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

