% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getWRF4sts}
\alias{getWRF4sts}
\title{Convert and extract WRF data at stations locations}
\usage{
getWRF4sts(filenames, st.loc, var4d = c("QKE"), var3d = c("SWDOWN", "PBLH"),
  vlev = 10, masl = FALSE)
}
\arguments{
\item{filenames}{a vector of WRF filenames (It can be length(filenames) == 1)}

\item{st.loc}{dataframe as produced by locWRFst()}

\item{vlev}{number of vertical levels (default = 10)}

\item{var4d}{vector of strings containing the name(s) of 4D WRF variable(s) to extract}

\item{var3d}{vector of strings containing the name(s) of 3D WRF variable(s) to extract}

\item{masl}{\code{logical} whether or not the elevation is above sea level}
}
\value{
list containing the dataframes for each station \cr
 c('name','lat','lon','date','u','v','w','ws','wd','p','T','theta','Td',var4d,var3d)
}
\description{
\code{getWRF4sts} extracts WRF default (meteorological variables) + additional variables at stations locations from WRF files.
}
\details{

}
\author{
I. Lopez-Coto, 2014 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

