% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getWRFcube}
\alias{getWRFcube}
\title{Convert WRF data to regular 3D cube (lon, lat, z)}
\usage{
getWRFcube(nc, time_ind, vlev = 59, HGT_WRF, lon_WRF, lat_WRF, WRFproj, zout,
  var4d = c("QKE", "tracer_10"), var3d = c("PBLH"))
}
\arguments{
\item{nc}{pointer to the WRF netcdf file as produced by \link{open.nc}}

\item{time_ind}{\code{integer} with the time index within the file being processed}

\item{vlev}{number of vertical levels (default = 10)}

\item{HGT_WRF}{3D terrain elevation array as produced by \link{getGRID}}

\item{lon_WRF}{2D WRF longitudes matrix as produced by \link{getGRID}}

\item{lat_WRF}{2D WRF latitudes matrix as produced by \link{getGRID}}

\item{WRFproj}{proj4 string witht the current projection as produced by \link{getGRID}}

\item{var4d}{vector of strings containing the name(s) of 4D WRF variable(s) to extract}

\item{var3d}{vector of strings containing the name(s) of 3D WRF variable(s) to extract}
}
\value{
list containing objects, mostly rasters, for the variables \cr
 c('U3D','V3D','W3D','WS3D','WD3D','P3D','T3D','TH3D','TD3D',var4d,var3d)
}
\description{
\code{getWRFcube} extracts WRF default (meteorological variables) + additional variables to regular 3D cube (lon, lat, z)  from WRF files.
 \strong{WARNING:} this function may be slow and require a lot of memory depending on the input file size.
}
\details{

}
\author{
I. Lopez-Coto, 2014 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

