% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getWRFprofile}
\alias{getWRFprofile}
\title{Convert and extract WRF vertical profiles at stations locations}
\usage{
getWRFprofile(filenames, st.loc, vlev = 50, masl = FALSE,
  var4d = c("QKE"))
}
\arguments{
\item{filenames}{a vector of WRF filenames (It can be length(filenames) == 1)}

\item{st.loc}{dataframe as produced by locWRFst()}

\item{vlev}{number of vertical levels (default = 10)}

\item{var4d}{vector of strings containing the name(s) of 4D WRF variable(s) to extract}

\item{masl}{\code{logical} whether or not the elevation is above sea level}
}
\value{
list containing the dataframes for each station \cr
 c('name','lat','lon','date','z','u','v','w','ws','wd','p','T','theta','Td',var4d)
}
\description{
\code{getWRFprofile} extracts WRF vertical profiles of default (meteorological variables) + additional variables at stations locations from WRF files.
}
\details{

}
\author{
I. Lopez-Coto, 2014 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

