% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ISDstations}
\alias{ISDstations}
\title{Retrieve ISD stations information}
\usage{
ISDstations()
}
\value{
data.frame with ISD stations list
}
\description{
\code{ISDstations} download the station list from the ftp://ftp.ncdc.noaa.gov/pub/data/noaa/ish-history.csv
}
\details{

}
\author{
I. Lopez-Coto, 2013 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

