% Generated by roxygen2 (4.0.1): do not edit by hand
\name{locWRFst}
\alias{locWRFst}
\title{Localize stations indexes}
\usage{
locWRFst(fileWRF, st.data)
}
\arguments{
\item{fileWRF}{a string containing the WRF filename}

\item{st.data}{a data.frame  [  st.data <- data.frame(st.data$Lat,st.data$Lon,st.data$AGL,st.data$Name)  ]}
}
\value{
A dataframe containing stations information \cr
 dimnames(st.loc)[[1]] = stations names # \cr
 dimnames(st.loc)[[2]] = c('lat','lon','elev','lat_indx','lon_indx')
}
\description{
\code{locWRFst} localizes the lat and lon indexes for each station within the WRF grid
}
\details{

}
\author{
I. Lopez-Coto, 2013 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

