% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qc}
\alias{qc}
\title{Write qc flag}
\usage{
qc(x, nodata = -888888, qcflag = 2^15)
}
\arguments{
\item{x}{observation}

\item{nodata}{no data value}

\item{qcflag}{qcflag value}
}
\value{
Write qc flag \cr
}
\description{
\code{qc} Write qc flag
}
\details{

}
\author{
I. Lopez-Coto, 2014 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

