% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readISD}
\alias{readISD}
\title{Read and reformat the ISD data file}
\usage{
readISD(filename, QC = TRUE)
}
\arguments{
\item{filename}{}

\item{QC}{\code{logical}}
}
\value{
data.frame with ISD station data
}
\description{
\code{readISD} read and reformat the ISD data file
}
\details{

}
\examples{
\dontrun{
 filename<-getISD()
 filename<-substring(filename,1,nchar(filename)-3)
 data <- readISD(filename,QC=FALSE)
 }
}
\author{
I. Lopez-Coto, 2013 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

