% Generated by roxygen2 (4.0.1): do not edit by hand
\name{slp2p}
\alias{slp2p}
\title{Sea Level Pressure to Pressure}
\usage{
slp2p(pseal, elev, Temp, n)
}
\arguments{
\item{pseal}{Pressures vector (hPa)}

\item{elev}{Station height above sea level (m)}

\item{Temp}{Temperature vector (C) (Use to be the 12 hours mean)}

\item{n}{\code{integer} number of elements to be used on the Temp running average}
}
\value{
pressure corrected to sea level
}
\description{
\code{slp2p} convert from pressure to sea level pressure
}
\details{

}
\author{
I. Lopez-Coto, 2013 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

