% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vert.interp}
\alias{vert.interp}
\title{Interpolate to fixed z coordinates}
\usage{
vert.interp(subset, zabg, zout)
}
\arguments{
\item{subset}{3D array with the variable to interpolate (lon, lat, vert)}

\item{zabg}{3D array with heights above ground for each grid cell (lon, lat, vert)}

\item{zout}{vector with the desired heights}
}
\value{
3D array with the interpolated field
}
\description{
\code{vert.interp} interpolate the heights to fixed z coordinatesn
}
\details{

}
\author{
I. Lopez-Coto, 2014 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

