% Generated by roxygen2 (4.0.1): do not edit by hand
\name{VM}
\alias{VM}
\title{Verification methods}
\usage{
VM(data.obs, data.fcst, plotMode = TRUE, main_info = NULL)
}
\arguments{
\item{data.obs}{vector of observations}

\item{data.fcst}{vector of forecast data}

\item{plotMode}{\code{logical}}

\item{main}{string to be used on the plots title}
}
\value{
data.frame
}
\description{
\code{VM} performes some verification methods
}
\details{

}
\author{
I. Lopez-Coto, 2012 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

