% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wrf2d.raster}
\alias{wrf2d.raster}
\title{Rasterize and Reproject 2D fields}
\usage{
wrf2d.raster(subset, lon_WRF, lat_WRF, proj.latlon, WRFproj, reproject = TRUE)
}
\arguments{
\item{subset}{a matrix with the input 2D field}

\item{lon_WRF}{a matrix with the WRF longitudes}

\item{lat_WRF}{a matrix with the WRF latitudes}

\item{proj.latlon}{proj4 string witht the desired projection}

\item{WRFproj}{proj4 string witht the current projection}

\item{reproject}{\code{logical} whether or not to reproject}
}
\value{
a reprojected raster object
}
\description{
\code{wrf2d.raster} reprojects and rasterizes WRF 2D fields to lat lon
}
\details{

}
\author{
I. Lopez-Coto, 2014 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

