% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wrf3d.brick}
\alias{wrf3d.brick}
\title{Rasterize and Reproject 3D fields}
\usage{
wrf3d.brick(subset3d, lon_WRF, lat_WRF, proj.latlon, WRFproj,
  reproject = TRUE)
}
\arguments{
\item{subset3D}{a 3D array with the input field}

\item{lon_WRF}{a matrix with the WRF longitudes}

\item{lat_WRF}{a matrix with the WRF latitudes}

\item{proj.latlon}{proj4 string witht the desired projection}

\item{WRFproj}{proj4 string witht the current projection}

\item{reproject}{\code{logical} whether or not to reproject}
}
\value{
a reprojected raster stack object
}
\description{
\code{wrf3d.brick} reprojects and rasterizes WRF 3D fields to lat lon
}
\details{

}
\author{
I. Lopez-Coto, 2014 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

