\name{wrfR}
\alias{wrfR}
\alias{wrfR}
\docType{package}
\title{
WRF post processing tools
-- README 
}
\description{
WRF post processing tools ...
}
\details{
\tabular{ll}{
Package: \tab wrfR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-07-30\cr
License: \tab GPL-3\cr
}
to be done...
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
I. Lopez-Coto (israel.lopez@dfa.uhu.es / inl@nist.gov)

Maintainer: I. Lopez-Coto (israel.lopez@dfa.uhu.es / inl@nist.gov)
}
\references{
- WRF references...
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ coming soon... ~~
}
