% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write.mat}
\alias{write.mat}
\title{Write fortran format}
\usage{
write.mat(m, codes, sep = "", ...)
}
\arguments{
\item{m}{matrix or data.frame}

\item{code}{a vector of codes}

\item{file}{file name string}

\item{...}{}
}
\value{
write formatted output\cr
}
\description{
\code{write.mat} write fortran formatted string
}
\details{

}
\author{
Gabor Grothendieck, 2010 (http://r.789695.n4.nabble.com/write-fortran-td1587119.html)
}

