% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write_obs}
\alias{write_obs}
\title{Write observations to littleR format}
\usage{
write_obs(xlat, xlon, string1 = NULL, string2 = NULL, string3 = NULL,
  string4 = NULL, ter, kx, mdate, slp, p, z, t, td, spd, dir, filename)
}
\arguments{
\item{xlat}{Latitude}

\item{xlon}{Longitude}

\item{string1}{Station description (Code and Name)}

\item{string2}{}

\item{string3}{Observation type ('FM-35 TEMP')}

\item{string4}{}

\item{ter}{Terrain elevation}

\item{kx}{Number of vertical levels}

\item{mdate}{Date}

\item{slp}{Sea level pressure (Pa)}

\item{p}{Pressures vector (Pa)}

\item{z}{Heights vector (m)}

\item{t}{Temperatures vector (K)}

\item{td}{Dew point vector (K)}

\item{spd}{Wind speed vector (m/s)}

\item{dir}{Wind direction vector (deg)}

\item{filename}{filename string}
}
\value{
write one observations report to littleR format\cr
}
\description{
\code{write_obs} write one observations report to littleR format. This function is based on the example upa.f
  originary from MM5 and currently distributed with obsgrid
}
\details{

}
\examples{
\dontrun{

kx<-10

p  <- c(1000.,   850.,   700.,   500.,   400.,   300., 250.,   200.,   150.,   100.)
z  <- c(100.,  1500.,  3000.,  5500.,  7000.,  9000., 10500., 12000., 13500., 16000.)
t  <- c( 14.,     6.,    -4.,   -21.,   -32.,   -45.,  -52.,   -57.,   -57.,   -57.)
td <- c( 13.,     3.,    -9.,   -28.,   -41.,   -55.,  -62.,   -67.,   -67.,   -67.)
spd<- c(  1.,     3.,     5.,     7.,     9.,    11., 13.,    15.,   17.,     19.)
dir<- c(  0.,    30.,    60.,    90.,   120.,   150.,  180.,   210.,  240.,    270.)

p=p*100.
t=t+273.15
td=td+273.15

slp <- c(101325.)
ter <- 1.
xlat <- 22.
xlon <- 115.
mdate <- '1995073018'

string1 <- '99001  Maybe more site info'
string2 <- 'SOUNDINGS FROM ????????? SOURCE'
string3 <- 'FM-35 TEMP'
string4 <- '    '

filename <- 'test.dat'


write_obs(xlat,xlon, string1,string2,string3,string4,ter,kx,mdate,
          slp, p, z, t,td,spd,dir,filename)
 }
}
\author{
I. Lopez-Coto, 2014 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

