% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xyz2raster}
\alias{xyz2raster}
\title{Rasterize xyz data to a regular lat lon grid}
\usage{
xyz2raster(data, subset, p.res = 100,
  proj.utm = "+proj=utm +zone=18 +units=km +datum=NAD83 +no_defs",
  log = TRUE)
}
\arguments{
\item{data}{data.frame(lon=lon,lat=lat)  ## All the stations coordinates}

\item{subset}{data.frame(lon=lon,lat=lat,conc=conc) ##  Subset of stations information}

\item{log}{\code{logical} whether or not to perform the fitting in log-scale}

\item{proj.utm}{proj4 string \cr
'+proj=utm +zone=31 +units=km +datum=WGS84 +no_defs' # ETEX  \cr
'+proj=utm +zone=18 +units=km +datum=NAD83 +no_defs' # ANATEX}
}
\value{
raster object with the "predicted" field
}
\description{
\code{xyz2raster}This function performs an ordinary kriging of the (log) concentration to "predict" the spatial distribution of the input value
}
\details{

}
\author{
I. Lopez-Coto, 2014 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

