% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getISD}
\alias{getISD}
\title{Retrieve the ISD data file}
\usage{
getISD(year = 2014, USAF = 84490, WBAN = 13025, stations = NULL)
}
\arguments{
\item{pathout}{}

\item{year}{}

\item{USAF}{}

\item{WBAN}{}
}
\value{
data.frame with ISD stations list
}
\description{
\code{getISD} download the station data from the ftp://ftp.ncdc.noaa.gov/pub/data/noaa/...
}
\details{

}
\author{
I. Lopez-Coto, 2013 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

