% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readISD}
\alias{readISD}
\title{Read and reformat the ISD data file}
\usage{
readISD(stdata, QC = TRUE)
}
\arguments{
\item{stdata}{character vector with one element per ISD raw report}

\item{QC}{\code{logical}}
}
\value{
data.frame with ISD station data
}
\description{
\code{readISD} read and reformat the ISD data file
}
\details{

}
\examples{
\dontrun{
 stdata<-getISD()
 data <- readISD(stdata,QC=FALSE)
 }
}
\author{
I. Lopez-Coto, 2013 (israel.lopez@dfa.uhu.es / inl@nist.gov)
}

