\name{wrfR}
\alias{wrfR}
\alias{wrfR}
\docType{package}
\title{
WRF post processing tools
-- README 
}
\description{

The \strong{R WRF post processing tools} is a suite of R tools developed to ease working in R with projected meteorological and air quality netcdf files (http://www.unidata.ucar.edu/software/netcdf/), mostly coming from WRF and WRF-CHEM (http://www.wrf-model.org) 

The current version v0.1 is a "raw" testing version developed on the base of some R functions I already had as a product of my research on Atmospheric Sciences. Please, be aware that, at this point, most of the functions need to be optimizated. 

This project aims to be collaborative and it is expected to grow with time as \strong{wrfseRs}  help to optimize, improve and extend this \strong{wrfR} package. 

\link{https://github.com/ilopezcoto/wrfR} \cr
\link{http://ilopezcoto.github.io/wrfR/}
}
\details{
\tabular{ll}{
Package: \tab wrfR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-07-30\cr
License: \tab GPL-3\cr
}
to be done...
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
I. Lopez-Coto (israel.lopez@dfa.uhu.es / inl@nist.gov)

Maintainer: I. Lopez-Coto (israel.lopez@dfa.uhu.es / inl@nist.gov)
}
\references{
- WRF references...
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ coming soon... ~~
}
